Splitting network and mqtt into separate interfaces and classes for Target and Sim HAL. Untested! Breaks current Apps!
