+++ lib/HALSim/src/Board.h
@@ -51,6 +51,7 @@
 #include "LedGreen.h"
 #include "LedRed.h"
 #include "Network.h"
+#include "MqttClient.h"
 
 /******************************************************************************
  * Macros
@@ -120,7 +121,7 @@ public:
             /* Log Device error */
             LOG_ERROR("Device process failed. ");
         }
-        else if (false == m_network.process())
+        else if (false == m_network.manageConnection())
         {
             /* Log Network error */
             LOG_ERROR("Network process failed. ");
